Add guard functions for switching app id
This is mostly a fix for devs, but if SDK check for valid tokens and checks them against server, but the “app id” has changed, then it’s an invalid situation. This code adds a check to require a match on “app id” (cid) as well or else forces a new SSO login.